+++ totalRP3/modules/dashboard/dashboard.lua
@@ -194,7 +194,7 @@ TRP3_API.dashboard.init = function()
 	end);
 
 	-- Tab bar
-	local whatsNewText = loc("WHATS_NEW_8");
+	local whatsNewText = loc("WHATS_NEW_9");
 	local moreModuleText = loc("MORE_MODULES_2");
 	local aboutText = loc("THANK_YOU_1");